refactor: clean up TreeView component
