+++ packages/dm-core/src/components/TreeView/index.ts
@@ -1,0 +1,2 @@
+export * from './TreeView'
+export * from './types'
