+++ packages/dm-core/src/components/TreeView/styles.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const StyledList = styled.ul`
+  padding: 0;
+  & & {
+    padding-left: 1rem;
+  }
+`
