+++ packages/dm-core/src/components/TreeView/types.ts
@@ -1,0 +1,7 @@
+import { TreeNode } from '../../domain/Tree'
+
+export type TNodeWrapperProps = {
+  node: TreeNode
+  setNodeOpen: (x: boolean) => void
+  children: any
+}
